refactor(dynamic unit): code simplifications
